[opt](inverted index) fix compile warning in compound reader

+++ be/src/olap/rowset/segment_v2/inverted_index_compound_reader.h
@@ -86,7 +86,7 @@ public:
                         int32_t read_buffer_size = CL_NS(store)::BufferedIndexInput::BUFFER_SIZE,
                         const io::IOContext* io_ctx = nullptr);
     ~DorisCompoundReader() override;
-    void copyFile(const char* file, int64_t file_length, uint8_t* buffer, int64_t buffer_length);
+    void copyFile(const char* file, int64_t file_length, uint8_t* buffer, int32_t buffer_length);
     bool list(std::vector<std::string>* names) const override;
     bool fileExists(const char* name) const override;
     int64_t fileModified(const char* name) const override;
